Move appendStepList into append.go
