+++ src/cmd/append.go
@@ -7,6 +7,7 @@ import (
 	"github.com/git-town/git-town/v7/src/dialog"
 	"github.com/git-town/git-town/v7/src/git"
 	"github.com/git-town/git-town/v7/src/runstate"
+	"github.com/git-town/git-town/v7/src/steps"
 	"github.com/spf13/cobra"
 )
 
@@ -109,4 +110,23 @@ func determineAppendConfig(args []string, repo *git.ProdRepo) (*appendConfig, er
 		shouldNewBranchPush: shouldNewBranchPush,
 		targetBranch:        targetBranch,
 	}, nil
+}
+
+func appendStepList(config *appendConfig, repo *git.ProdRepo) (runstate.StepList, error) {
+	result := runstate.StepList{}
+	for _, branch := range append(config.ancestorBranches, config.parentBranch) {
+		steps, err := syncBranchSteps(branch, true, repo)
+		if err != nil {
+			return runstate.StepList{}, err
+		}
+		result.AppendList(steps)
+	}
+	result.Append(&steps.CreateBranchStep{Branch: config.targetBranch, StartingPoint: config.parentBranch})
+	result.Append(&steps.SetParentBranchStep{Branch: config.targetBranch, ParentBranch: config.parentBranch})
+	result.Append(&steps.CheckoutBranchStep{Branch: config.targetBranch})
+	if config.hasOrigin && config.shouldNewBranchPush && !config.isOffline {
+		result.Append(&steps.CreateTrackingBranchStep{Branch: config.targetBranch, NoPushHook: config.noPushHook})
+	}
+	err := result.Wrap(runstate.WrapOptions{RunInGitRoot: true, StashOpenChanges: true}, repo)
+	return result, err
 }
