+++ src/cmd/core.go
@@ -28,7 +28,6 @@ import (
 	"github.com/git-town/git-town/v7/src/git"
 	"github.com/git-town/git-town/v7/src/hosting"
 	"github.com/git-town/git-town/v7/src/runstate"
-	"github.com/git-town/git-town/v7/src/steps"
 )
 
 func validateIsConfigured(repo *git.ProdRepo) error {
@@ -46,25 +45,6 @@ func ValidateIsRepository(repo *git.ProdRepo) error {
 		return errors.New("this is not a Git repository")
 	}
 	return repo.NavigateToRootIfNecessary()
-}
-
-func appendStepList(config *appendConfig, repo *git.ProdRepo) (runstate.StepList, error) {
-	result := runstate.StepList{}
-	for _, branch := range append(config.ancestorBranches, config.parentBranch) {
-		steps, err := syncBranchSteps(branch, true, repo)
-		if err != nil {
-			return runstate.StepList{}, err
-		}
-		result.AppendList(steps)
-	}
-	result.Append(&steps.CreateBranchStep{Branch: config.targetBranch, StartingPoint: config.parentBranch})
-	result.Append(&steps.SetParentBranchStep{Branch: config.targetBranch, ParentBranch: config.parentBranch})
-	result.Append(&steps.CheckoutBranchStep{Branch: config.targetBranch})
-	if config.hasOrigin && config.shouldNewBranchPush && !config.isOffline {
-		result.Append(&steps.CreateTrackingBranchStep{Branch: config.targetBranch, NoPushHook: config.noPushHook})
-	}
-	err := result.Wrap(runstate.WrapOptions{RunInGitRoot: true, StashOpenChanges: true}, repo)
-	return result, err
 }
 
 // handleUnfinishedState checks for unfinished state on disk, handles it, and signals whether to continue execution of the originally intended steps.
